backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client (for AI API calls)
httpx==0.26.0

# AI Providers (candidates can use any of these)
openai==1.10.0
anthropic==0.18.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
